move ch22 to ch23, ch21 to ch22
