+++ docs/chapter_blurbs.md
@@ -24,7 +24,7 @@ What does each one do?
 - **ch18_world_etl**: Defines the tools that use Idea Bricks to build MomentUnits.
 - **ch19_world_kpi**: Defines the analytics outcomes of completed MomentUnits.
 - **ch20_world_apex**: Defines WorldDirs that admin MomentUnits input data and outcomes
-- **ch21_lobby**: Defines Lobby Tools for gifts, World Scenarios.
-- **ch22_finance**: Defines Finance Tools for 
+- **ch22_lobby**: Defines Lobby Tools for gifts, World Scenarios.
+- **ch23_finance**: Defines Finance Tools for 
 - **ch24_person_viewer**: Tools for Visualizing PersonUnits
 - **ch98_docs_builder**: Defines Tools that create documentation.
