--- src/ch22_finance/test/_util/test_ch22_path.py
@@ -1,13 +1,0 @@
-from inspect import getdoc as inspect_getdoc
-from platform import system as platform_system
-from src.ch00_py.file_toolbox import create_path
-from src.ch21_lobby._ref.ch21_path import (
-    LobbyID,
-    create_lobby_dir_path,
-    create_moment_mstr_dir_path,
-    create_world_dir_path,
-)
-from src.ch22_finance.test._util.ch22_env import get_temp_dir
-from src.ref.keywords import Ch22Keywords as kw
-
-# financial tooling
