+++ src/ch22_lobby/_ref/ch22_ref.json
@@ -1,6 +1,6 @@
 {
   "chapter_blurb": "Defines Lobby Tools for gifts, World Scenarios.", 
   "chapter_content": "Defines Lobby Tools for gifts, World Scenarios.", 
-  "chapter_description": "ch21_lobby", 
-  "chapter_number": 21
+  "chapter_description": "ch22_lobby", 
+  "chapter_number": 22
 }
