+++ src/ch22_lobby/test/_util/test_ch22_path.py
@@ -1,14 +1,14 @@
 from inspect import getdoc as inspect_getdoc
 from platform import system as platform_system
 from src.ch00_py.file_toolbox import create_path
-from src.ch21_lobby._ref.ch21_path import (
+from src.ch22_lobby._ref.ch22_path import (
     LobbyID,
     create_lobby_dir_path,
     create_moment_mstr_dir_path,
     create_world_dir_path,
 )
-from src.ch21_lobby.test._util.ch21_env import get_temp_dir
-from src.ref.keywords import Ch21Keywords as kw
+from src.ch22_lobby.test._util.ch22_env import get_temp_dir
+from src.ref.keywords import Ch22Keywords as kw
 
 
 def test_LobbyID_Exists():
