+++ src/ch23_finance/_ref/ch23_path.py
