+++ src/ch23_finance/_ref/ch23_ref.json
@@ -1,6 +1,6 @@
 {
   "chapter_blurb": "Defines Finance Tools for ", 
   "chapter_content": "Defines Finance Tools for financial modeling", 
-  "chapter_description": "ch22_finance", 
-  "chapter_number": 22
+  "chapter_description": "ch23_finance", 
+  "chapter_number": 23
 }
