+++ src/ch23_finance/_ref/ch23_semantic_types.py
@@ -25,4 +25,4 @@ from src.ch11_bud._ref.ch11_semantic_types import SparkInt, TimeNum
 from src.ch12_keep._ref.ch12_semantic_types import ManaNum
 from src.ch13_time._ref.ch13_semantic_types import EpochLabel
 from src.ch20_world_apex._ref.ch20_semantic_types import WorldName
-from src.ch21_lobby._ref.ch21_semantic_types import LobbyID
+from src.ch22_lobby._ref.ch22_semantic_types import LobbyID
