+++ src/ch23_finance/test/_util/test_ch23_path.py
@@ -1,0 +1,7 @@
+from inspect import getdoc as inspect_getdoc
+from platform import system as platform_system
+from src.ch00_py.file_toolbox import create_path
+from src.ch23_finance.test._util.ch23_env import get_temp_dir
+from src.ref.keywords import Ch23Keywords as kw
+
+# financial tooling
