+++ src/ref/keywords_main.json
@@ -23,7 +23,7 @@
   "KnotTerm": {"init_chapter": "ch04", "semantic_type": "str"}, 
   "LabelTerm": {"init_chapter": "ch01", "semantic_type": "str"}, 
   "Lasso": {"init_chapter": "ch09"}, 
-  "LobbyID": {"init_chapter": "ch21", "semantic_type": "str"}, 
+  "LobbyID": {"init_chapter": "ch22", "semantic_type": "str"}, 
   "ManaGrain": {"init_chapter": "ch04", "semantic_type": "float"}, 
   "ManaNum": {"init_chapter": "ch12", "semantic_type": "float"}, 
   "MomentRope": {"init_chapter": "ch09", "semantic_type": "str"}, 
@@ -151,8 +151,8 @@
   "factheirs": {"init_chapter": "ch05"}, 
   "factunits": {"init_chapter": "ch05"}, 
   "filename": {"init_chapter": "ch00"}, 
-  "finance": {"init_chapter": "ch22"}, 
-  "financial": {"init_chapter": "ch22"}, 
+  "finance": {"init_chapter": "ch23"}, 
+  "financial": {"init_chapter": "ch23"}, 
   "five": {"init_chapter": "ch13"}, 
   "found_facts": {"init_chapter": "ch11"}, 
   "fund_agenda_give": {"init_chapter": "ch02"}, 
@@ -216,9 +216,9 @@
   "laborunit": {"init_chapter": "ch02"}, 
   "lasso": {"init_chapter": "ch09"}, 
   "last_lesson_id": {"init_chapter": "ch07"}, 
-  "lobby_id": {"init_chapter": "ch21"}, 
-  "lobby_mstr_dir": {"init_chapter": "ch21"}, 
-  "lobbys": {"init_chapter": "ch21"}, 
+  "lobby_id": {"init_chapter": "ch22"}, 
+  "lobby_mstr_dir": {"init_chapter": "ch22"}, 
+  "lobbys": {"init_chapter": "ch22"}, 
   "magnitude": {"init_chapter": "ch11"}, 
   "mana_grain": {"init_chapter": "ch04"}, 
   "mandate": {"init_chapter": "ch07"},
